Moved app.js from src folder to the root of the project, and changed the path to app in server.js
