+++ server.js
@@ -1,4 +1,4 @@
-const app = require("./src/app");
+const app = require("./app");
 const port = process.env.PORT || 5000;
 
 app.set("port", port);
